requirements.txt
torch
transformers
clip-by-openai
faiss-cpu
matplotlib
Pillow
tqdm
numpy